requirements.txt
streamlit==1.50.0
tensorflow-cpu==2.19.0
pandas==2.3.3
numpy==2.0.2
scikit-learn==1.6.1
plotly==6.3.1
matplotlib==3.9.4
shap==0.48.0
reportlab==4.4.4
seaborn==0.13.2
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
protobuf==3.20.3